+++ app/.firebase/hosting.ZGlzdA.cache
@@ -1,20 +1,20 @@
-index.html,1696025001958,968b9d727b96f33828a0ef7014ca149f6fc78d155858c25d9b97ee538990f87f
-vite.svg,1696025001522,74d16a12383b458298b68f07379e888b23e901da036b1d53ece0675b95b90cf2
-roxana.ico,1696025001521,76d9e19334b90e899725781eb2b29959f196c9516347c1473110c022a02cfeb7
-assets/humanComputerInteraction-4e4ed418.js,1696025001958,0ab6bd1b63e3904d45bce96f15bbb85eaff0740915d6ce7388c63a5b9eed1642
-assets/focusInvesting-d019fb98.js,1696025001958,2f67c62e911335a5282d3e2ad7ae6addb11e7c7fc97124f159465ee6af86bbb6
-assets/selfEsteemAddendum-a6d06cac.js,1696025001958,b56d9fd65d3adff495735e6a84ddd01e1e58aa35bc97e307c009f6419a24de98
-assets/linkedInLogo-3c0149f2.png,1696025001958,fe698056cdc7d5225ffb0adf90acb88e6935979f60964235cf5c3cdf659a63f8
-assets/wealthManagement-70a9c728.js,1696025001958,ba05f2fb03802ef5966d23ebd988305a2e0029541b2f78a642cb69ac81078f9f
-assets/philosophy-df63fed3.js,1696025001958,d2fe1e2ffbc002c8fc039931c0fdd7d116999d3b512976049f782aba614fa7b9
-assets/khanac-a717d451.jpeg,1696025001958,4000bac1d48161404973cb789f9261b6ac63dcb4eb0a5af9f94717151b34deec
-assets/phillyd-8d33f14a.jpeg,1696025001957,199dd8b1e41382fef700da75a0c67a3fc0aa39d625f4578ec80ba8d85a6824b2
-assets/patreonLogos-eab0f4b9.png,1696025001958,8a3157f8186ad653f735b323e06a44b5afe6d76e2a147f716205ce53642e38de
-assets/snow-533c45a5.jpeg,1696025001958,e05ec11245a6daddb692fe79c25e5e21899e9885dc25fc2cc9c6f748d08c99e8
-snowthaproduct.jpeg,1696025001522,cf9a08eb3a5aecf55007b4e966b4ac7c8303f3557b54ee692916e06c38abc6b1
-assets/index-1b997570.css,1696025001958,12414fa931ec1bdb625f30988417df84b615bbbbc62079f7d4c78be0089a35a7
-assets/cashAppCard-be953d39.jpeg,1696025001951,abc4ce70e6d8683b4496b1124ec117057f990273a373d772a11fbab6f2b055ab
-assets/sheilferBitcoin-7793fc9a.jpeg,1696025001952,1715d00e76e5c8a9e112b699d34604bc3eb7fa3c4bba3d26a0526957405bb50f
-assets/roxanabrand-9779279c.gif,1696025001957,8289e6e613ee57e385c5a0953732e70787b021b64e92231aea40450129732fb3
-assets/roxanaGif-0e1f9208.gif,1696025001958,132f360e6f71fcb5f473002f5b5fc5a8db6ae621291546a12b3ff8f8d6561d8e
-assets/index-f57f381f.js,1696025001959,4451716cde8c045d6293cf412f95bd97a2cbe6d875e3873b8d443a06fa7f7291
+index.html,1696025650126,dc31aae8a190531df3cd21585cf16212569c65343572f049c9f1058aa706633a
+vite.svg,1696025649699,74d16a12383b458298b68f07379e888b23e901da036b1d53ece0675b95b90cf2
+roxana.ico,1696025649699,76d9e19334b90e899725781eb2b29959f196c9516347c1473110c022a02cfeb7
+assets/humanComputerInteraction-4e4ed418.js,1696025650125,0ab6bd1b63e3904d45bce96f15bbb85eaff0740915d6ce7388c63a5b9eed1642
+assets/linkedInLogo-3c0149f2.png,1696025650126,fe698056cdc7d5225ffb0adf90acb88e6935979f60964235cf5c3cdf659a63f8
+assets/focusInvesting-d019fb98.js,1696025650126,2f67c62e911335a5282d3e2ad7ae6addb11e7c7fc97124f159465ee6af86bbb6
+assets/selfEsteemAddendum-a6d06cac.js,1696025650125,b56d9fd65d3adff495735e6a84ddd01e1e58aa35bc97e307c009f6419a24de98
+assets/philosophy-df63fed3.js,1696025650126,d2fe1e2ffbc002c8fc039931c0fdd7d116999d3b512976049f782aba614fa7b9
+assets/wealthManagement-70a9c728.js,1696025650125,ba05f2fb03802ef5966d23ebd988305a2e0029541b2f78a642cb69ac81078f9f
+assets/khanac-a717d451.jpeg,1696025650125,4000bac1d48161404973cb789f9261b6ac63dcb4eb0a5af9f94717151b34deec
+assets/phillyd-8d33f14a.jpeg,1696025650125,199dd8b1e41382fef700da75a0c67a3fc0aa39d625f4578ec80ba8d85a6824b2
+assets/patreonLogos-eab0f4b9.png,1696025650125,8a3157f8186ad653f735b323e06a44b5afe6d76e2a147f716205ce53642e38de
+assets/snow-533c45a5.jpeg,1696025650125,e05ec11245a6daddb692fe79c25e5e21899e9885dc25fc2cc9c6f748d08c99e8
+snowthaproduct.jpeg,1696025649699,cf9a08eb3a5aecf55007b4e966b4ac7c8303f3557b54ee692916e06c38abc6b1
+assets/index-1b997570.css,1696025650126,12414fa931ec1bdb625f30988417df84b615bbbbc62079f7d4c78be0089a35a7
+assets/cashAppCard-be953d39.jpeg,1696025650125,abc4ce70e6d8683b4496b1124ec117057f990273a373d772a11fbab6f2b055ab
+assets/sheilferBitcoin-7793fc9a.jpeg,1696025650126,1715d00e76e5c8a9e112b699d34604bc3eb7fa3c4bba3d26a0526957405bb50f
+assets/roxanabrand-9779279c.gif,1696025650125,8289e6e613ee57e385c5a0953732e70787b021b64e92231aea40450129732fb3
+assets/roxanaGif-0e1f9208.gif,1696025650126,132f360e6f71fcb5f473002f5b5fc5a8db6ae621291546a12b3ff8f8d6561d8e
+assets/index-74078ee6.js,1696025650127,39f618fbc779c1dd3bf1ce267296707c96bccc84bd203e00bec23e994a0af54a
