--- app/src/Analytics/Analytics.ts
@@ -1,2 +1,0 @@
-// handles analytics in one place. can be a class
-export const runAnalytics = () => {};
